Added basic scene management
Created a new manager scene (May be wrong) and set it as default. Added a script to deal with switching between scenes. This is a rough sketch, it will definitely need to be revised once we have added in scenes. Perhaps I should save this to the master, but I don't want to break anything - Jacob