Work so far
I am checking to see if the objects are null, removing them if I need to, etc. It doesn't work!!!!!!!!! One would think it would, but it doesn't. I'm a little peeved off right now, because I just don't get it. Is the turn off? Why does it crash when I am trying to see if something isn't null. Should it do that? I don't know