Added A lot of stuff and intruduced a lot of bugs
I added a player leveling system, looped audio, and player health mechanics thanks to Jacob's code. I also tried to add the mana system but for some reason it isn't working correctly. I added a lot of things to the saving/loading system and broke that but it isn't fully implemented yet. I also changed how the player is loaded into the game by toggling the mesh renderer and is kinematic on the rigidbody and introduced some big problems.