Messed with a navmesh on a mesh collider
Found some issues with the way we were doing navmeshes. I couldn't put a flat navmesh on the forest level because it is too bumpy. Also had problems with the navMeshAgent getting stuck on obsticles in the forest level. Fixed it by stopping the path when the player is moving less than 1 mps for one second.