General fixing of stuff
I got rid of the credits completely, as they were still loading in the background I think, which is bad. I also fixed a few random bugs, including the spell not working if it was the same number as the cost, and the menus being wack. I think this is the homestretch guys. The only things I can think of that we need to do is
-Fix the enemy hit audio, currently it is set as the death sound so there is something there.
-Fix some balancing. The game is impossible if you never upgrade melee, but too easy if that is all you do.
-There is a bug where when you level up on a level, but die later on in that level, you keep your earned level  when you respawn.